pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Carlos <carlos_pedro17@live.com>"]
readme = "README.md"
packages = [{include = "bookstore"}]

[tool.poetry.dependencies]
python = ">=3.13"
django = ">=5.2.6,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
django-extensions = ">=4.1,<5.0"
django-debug-toolbar = ">=6.0.0,<7.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
factory-boy = ">=3.3.3,<4.0.0"
pytest = ">=8.4.2,<9.0.0"
whitenoise = "^6.11.0"
gunicorn = "^23.0.0"
gitpython = "^3.1.45"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
